pyproject.toml
[tool.poetry]
name = "zorya"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
google-auth = "^1.30.0"
google-api-python-client = "^2.3.0"
google-cloud-pubsub = "^2.4.1"
google-cloud-logging = "^2.3.1"
google-cloud-firestore = "^2.1.1"
pytz = "^2021.1"
backoff = "1.10.0"
numpy = "^1.20.2"
gunicorn = "^20.1.0"
pydantic = "^1.8.1"
fastapi = "^0.64.0"
aiofiles = "^0.6.0"
uvicorn = {extras = ["standard"], version = "^0.12.0"}
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.1"
pytest = "^6.2.4"

[tool.poetry.scripts]
zorya = 'zorya.cli.cli:cli'

[tool.black]
line-length = 79

[tool.coverage.run]
omit = [
    "zorya/tests/*"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
